requirements.txt
Pillow==6.2.2
streamlit==1.18.1
